requirements.txt
PyQt6
yt-dlp
python-vlc
requests
